+++ tests/test_auth.py
@@ -1,6 +1,6 @@
-import os
 import unittest
 
+from tests.utils import integration_test
 from wavyfm import WavyClientCredentials, WavyAuthError
 
 try:
@@ -10,16 +10,6 @@ except ImportError:
 
 patch = mock.patch
 DEFAULT = mock.DEFAULT
-
-
-def integration_test(f):
-    def wrapper(self, *args, **kwargs):
-        if os.getenv("WAVYFM_RUN_INTEGRATION_TESTS", "0") == "1":
-            f(self, *args, **kwargs)
-        else:
-            self.skipTest("Integration tests disabled; enable with WAVYFM_RUN_INTEGRATION_TESTS=1")
-
-    return wrapper
 
 
 class TestWavyClientCredentials(unittest.TestCase):